+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add support of Chrome driver v115+.
+  Fixes issue: [#14 Webdriver can't be downloaded for new chrome ver. 115 and higher](https://github.com/atata-framework/atata-webdriversetup/issues/14).
+
 ## [2.6.0] - 2023-05-02
 
 ### Added